Refactors retry test by extracting FakeRetryTaskFn into dedicated file
